[Refactor] Lazy-loaded tool_parser
